Ctrate ttl index for tokens; Store refresh token; Implement logout

+++ bff/src/dtos/general_res_dto.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 #[derive(Debug, Deserialize, Serialize)]
 pub struct GeneralResDto {
-    pub status_code: u32,
+    pub status_code: u16,
     pub message: String
 }
